chatbot/requirements.txt
SQLAlchemy
python-dotenv
fastapi
uvicorn
langchain-openai
langchain-community
langchain-core
langchain-text-splitters
python-docx
pywebview
pyinstaller
pydantic
duckduckgo-search
pandas
faiss-cpu
arxiv